Launcher3: Do not attach finish() to frame update event
When the animation get canceled, finish() should be run instantly
because it does something like removing pending animations. The
switchToScreenshot() attaches it to RtFrameCallback which is only called
on frame update. This is leading to problems as the next frame update
may only happen when the next gesture starts, which means all the
pending animations will be removed on next gesture. And then the next
gesture is freezed.

Fixes random gesture freeze since 12L.

Change-Id: I10247294a2dcae467706c434685b299f8b525888
Signed-off-by: LibXZR <i@xzr.moe>