Update workspace items created via drag-and-drop in place.
When a URI is dropped on launcher, we create a new workspace item for
that URI. We then move the backing file associated with the dropped URI
to the "Home screen/" directory on the file system using media store
APIs. This triggers an observable event which we should handle to update
the created workspace item.

Currently, we create a new workspace item rather than updating the
created workspace item in place. This occurs because we need to be able
reconcile the created workspace item with its associated media store
URI but were previously unable to do so. This CL accomplishes that by
caching URI aliases that can then be used during update tasks.

Note that this CL also migrates the base URI that we operate on from
VOLUME_EXTERNAL to VOLUME_EXTERNAL_PRIMARY. This is because the URIs
that we receive from MediaStore#getMediaURI() use the latter, so we
need to observe the latter in order to resolve URI aliases. This may
need to change in the future but should be sufficient for now.

Demo (before): http://dr/file/d/1YjCaDbSmC7xKPs4EKxCIEKbXvnjek5k1/view
Demo (after): http://dr/file/d/1_LOzzcmEadOFX2NEN6fp7YRz7uEk7XvP/view
Design: go/system-drag-and-drop-of-files-in-launcher

Bug: 442049644
Flag: com.android.launcher3.enable_system_drag
Flag: com.android.launcher3.show_files_on_home_screen
Test: atest HomeScreenFilesChangedTaskTest
Test: atest HomeScreenFilesProviderTest
Test: atest SystemDragIntegrationTest
Change-Id: I388661ecb336fecb2b12ed7721d3abb24cab050e