Clean-up widget resize frame transition flag.
This is to simplify the class for further refactoring. I tried to turn
on this flag and it would need more alignments to make it fully
functional. So, we when take up this task, we can re-look how to
implement it and can always reference previous implementation if
necessary from bug / code search.

Bug: 268553314
Flag: EXEMPT FLAG_REMOVAL
Test: Manual regression test to validate resize frame functioning.
Change-Id: I8c2e18429a521e4e8009ecf9a9aecd1919298c04