Upon recents cancel generalize finishedToApp check
When a START_RECENTS transition is animating,
and the animation gets cancelled for some reason -
e.g. a transition merge() comes in and requests
a jump-cut cancel of recents animation - we need
to evaluate whether we are finishedToApp.

This should lead to proper state advancements
in TaskbarLauncherStateController and ultimately
determine whether the taskbar is stashed/unstashed
or visible at all.

Bug: 441656288
Flag: EXEMPT BUGFIX
Test: manually repro the steps in the bug
Change-Id: I4dda3031c9c4aaa5713c9a2c0634abb219fa0851