Enforce permission check for accepting system level drop.
When determining whether to accept a system level drop, we need to
verify that permissions were successfully obtained. Failure to do so
could result in a bad actor piggy-backing on the permissions that the
Launcher already has in order to move files on the file system.

Note that this CL also improves the SystemDragIntegrationTest suite to
perform an actual drag-and-drop sequence rather than a synthetic one.

Bug: 445710341
Bug: 445710944
Flag: com.android.launcher3.enable_system_drag
Test: atest SystemDragControllerImplTest SystemDragIntegrationTest
Change-Id: I12a06eaf36c2922257c6b56fa9370e3d2644a716