pin-overflow: Show error message if pinning 17th app
Currently we restrict users to only be able to pin 16 apps in the
taskbar. This is mainly to prevent users from overloading the
taskbar and can also reduce the engineering work.
As a guidance, we'll show an error message to users to notify that
there is no more space in the taskbar to pin an app.

Test: Manual && PinToTaskbarShortcutTest
Flag: com.android.window.flags.enable_overflow_button_for_taskbar_pinned_items
Bug: 375647992
Bug: 445827933
Change-Id: Ia43a9994a157186236f4bd683be5fba699bf65b6