Consume touch event when we dimiss popup menu
On widgets we were not consuming the touch event when a popup menu was
open. We now make sure to consume it so the user doesn't accidentally
launch other apps when trying to dismiss a popup. This follows the same
behavior as other item types when dismissing a popup.

Bug: 445991327
Test: manual
Flag: com.android.launcher3.home_screen_edit_improvements
Change-Id: Ia12de716fb97a656725f562182a2d4788dee51b2