Reserve space for search container when inserting home screen file.
This CL is analogous to ag/35443829 in that it reserves space for the
search container when inserting new home screen files into the layout/
model. Note that this is not an issue when injectable model items are
enabled, though that feature flag may not be kept in perfect sync with
that of the home screen files feature and so could result in user
confusion without this patch.

This CL introduces code which is a bit duplicative but it should be
short-lived and allows us to keep code changes behind the existing
home screen files feature flag w/o modifying lower level code.

Bug: 442874353
Flag: com.android.launcher3.show_files_on_home_screen
Test: atest HomeScreenFilesChangedTaskTest
Change-Id: I88d049fe957a9795407be789e266ca3f4e396db3