pin-overflow: Fix the item can't be pinned when overflow is shown
After ag/35510895, there is a bug where the new apps can not be
pinned properly from the Launcher All apps when the overflow icon
is shown. This is because the original way to check if an item is
pinned only checks the "hotseat" where the overflow only happens
in "taskbar". This cl fixes that by iterating over the overflow
apps as well as the apps on the hotseat.

Test: Manual
Flag: com.android.window.flags.enable_overflow_button_for_taskbar_pinned_items
Bug: 375647992
Bug: 446611357
Bug: 447207396
Change-Id: I8cd19184df3f9560ca67dc87b7ad7b12a0262b54