Show close popup option in desktop taskbar on home
Add close option for taskbar icons whenever taskbar is showing desktop
tasks, even if desktop mode is not technically active - e.g. on home
screen on desktop devices. Keep the logic where the option is excluded
in overview due to concerns with UX.

Additionally, removes close option if the app has no
recent/running/minimized tasks. Updates the action to operate on desktop
tasks only, to be consistent with the app running state displayed in the
taskbar.

Bug: 447091985
Bug: 445789590
Test: Verify that close option in taskbar popup menu shows up in
desktop-first mode on home, and that it doesn't show up if the app has
no open tasks.
Flag: EXEMPT BUGFIX

Change-Id: Ic670932c4b667fd3e3ebd601be467d6a030b6047