[Taskbar][10/n] Allow adding RecentsAnimationListener from taskbar ui thread
This CL re-implements ag/35613258 that was reverted: because caller has
complex listener register and unregister logic, it's not feasible to
return a SafeClosable and let caller remember which one to close.

Instead this CL let TaskbarLauncherStateController register a listener
that can be executed on taskbar ui thread.

Bug: 404636836
Flag: com.android.launcher3.enable_taskbar_ui_thread
Test: manual, tested animation of open apps from recents works
Change-Id: Ia490d3c0e851c1958c584ead156e54b952c4a29b