Temporary fix for race condition during home screen files init.
The initial query for home screen files occurs very early in both the
user session and application lifecycles. This can result in the query
failing if the external storage directory hasn't yet mounted.

This CL slaps a bandaid on the problem by giving the storage directory
500ms to mount. Note that this has the potential to block the loader
task thread, though the clock starts when the loader task is created,
not when it is run. This bandaid serves only to unblock dogfooding of
the home screen files feature on desktop but will be fast followed with
a more robust solution that does not block the loader task after
derisking the dogfood timeline.

Bug: 444563784
Flag: com.android.launcher3.show_files_on_home_screen
Test: atest HomeScreenFilesProviderTest
Change-Id: I10160b56b2285f75dd60032cb3e148cafef8b822