Implement onDrop for pinned/unpinned drop targets
- Pin a recent app if the user drags it from the recent apps on the
taskbar and drop it to the hotseats.
- Handle the cases when the user drags it from the hotseats and drop
it to the recent apps area: Unpins it if the app is opened and removes
it from the pinned apps area if it the app is not opened.

Screencast:
go/screenshot-ndg3ntqznzy2otg3ntcxmnwwzdy2oddlzc1hmq

Bug:447444838
Test: verified on test devices
Flag: com.android.launcher3.enable_taskbar_drag_and_drop
Change-Id: Ida321565e0be35bd2e234e093a05536e5ac43c68