Prevent settling on BackgroundApp state in onTransitionCancelled
BaseContainerInterface.onTransitionCancelled is expected to run during a quick switch from an app, which will leave the state manager in the BackgroundApp state. This is fine unless there is a configuration change that prevents StatefulActivity.onStop from moving the state back to Normal. So if the user happens to go home using a back gesture, then they will be stuck in the BackgroundApp state. Adding fallback logic to handle all similar series of events.

Flag: EXEMPT BUGFIX
Fixes: 433549260
Test: manual gestures with logs
Change-Id: If309814a660dfa3e166f930c65bde6dddeb5e2f3