Launcher Test Week OseWidgetViewTest
Moving OseWidgetViewTest to multivalent. Removed dependency on ExtendedMockito and only use spy().
In this migration, closeActionList was having "zero interactions" error
due to the test verifying a mock RunnableList that the OseWidgetView spy
is never actually using. When mVut = spy(OseWidgetView(spiedContext) is
used, the OseWidgetView object is initialized with a real RunnableList
insance. The fix here is to replace the real RunnableList instance
inside the spy with the mock instance.

Another issue in this migration is that the
WidgetUtils.findWidgetProvider() was returning an NPE. To fix this a
providerInfo is created whenever it is not found. This should be fine
since WidgetUtils#findWidgetProvider() only happens in test.

bug: 449753776
Test: atest NexusLauncherRoboTests:com.android.launcher3.qsb.OseWidgetViewTest
Flag: EXEMPT TEST_ONLY
Change-Id: I3bb1e2493518a5b3793d026e46a5eedf9738f64d