Update the icon cache with the new icon and title
`IconCache#getTitleAndIcon` first checks the in-memory cache, applying
the cached bitmap and title if found, which resulted in showing the old
icon or title after renaming the item in the file manager.

Additionally, updates to the title were not propagated to the launcher
db, causing the old title to be displayed after reboot.

Bug: 449860858
Bug: 449843153
Flag: com.android.launcher3.show_files_on_home_screen
Test: com.android.launcher3.homescreenfiles.HomeScreenFilesChangedTaskTest
Change-Id: I873c51db90c37bef970afa77510ae1247eedf25d