Mark AppWidgetResizeFrame closed before removing it
When handling close for `AppWidgetResizeFrame`, update `mIsOpened` value
to false before removing the veiw from the hierarchy. This ensures that
the AppWidgetResizeFrame view is marked as closed when it gets removed.
Without this, BaseDragLayer will think that the AppWidgetResizeFrame
view is still open when it notices the view removal, and asynchronously
schedule another close.
One effect is that the current widget view's location in the workspace
would be asynncronously be marked as occupied, even though the widget
may get removed in the meantime - for example, resize frame will be
shown with widget popup menu, and clicking on the remove option in the
menu will close the frame before passing on the event to the popup. If
close is scheduled to run asynchronously again, it's possible for the
repeated close runs after the widget's already been removed.

Bug: 450525011
Test: Place a widget in the first available slot in the workspace.
      Drag the  widget so resize frame shows up, right click to show
      popup menu, click remove option in the popup. Add the same widget
      using "Add" button. Verify the widget gets added in the first
      available slot (i.e. the one previously occupied by the removed
      widget)
Flag: EXEMPT BUGFIX

Change-Id: I0d8a8a772ad6c3e09477b746ea8cda566ef5d8ab