Replace ENABLE_TASK_CONTINUITY with TASK_CONTINUITY in Launcher Code
The `enable_task_continuity` flag was created under the `desktop_better_together` namespace back in April. Sometime in July, changes to the Desktop release process were made. All flags with a `desktop_*` namespace are forced under the "Desktop" feature flag rollout, rather than the "Platform" one.

Handoff is intended to be a platform feature, and thus this will not work. Because there is no way to change a flag's workspace, the only path forward is to create a new flag (`TASK_CONTINUITY`) under a platform namespace (`companion`), and replace all existing references with this one.

Flag: android.companion.task_continuity
Bug: 400970610
Test: Updated Tests

Change-Id: Ia2e273cde2102f51db6561a2a7a35a264ec1e0c7