Do not reuse previous recent icons for new ones in LTR.
This approach does not work for RTL UI right now, and we are replacing
this code with a LinearLayout container anyway. It will be easier to
support both layout directions with that approach. But this change will
mitigate binding performance for LTR.

Test: Manual, TaskbarViewTests
Bug: 315354060
Flag: com.android.launcher3.enable_recents_in_taskbar
Change-Id: I79053336c93c8c653f4a8a337f3caa67022bd4b3