Reuse existing GroupTask instances when non-desktop recents update.
When we do #getTasks, we get a new set of GroupTask instances, even for
ones we had previously. For Desktop running apps, we reuse the previous
instances, and then appand the brand new tasks. In contrast, we were
using the new instances for non-desktop Recents. The problem is we are
storing BitmapInfo in GroupTask, which means we will lose this data when
reloading tasks; we don't refetch BitmapInfo if the tasks did not
change.

Flag: com.android.launcher3.enable_recents_in_taskbar
Bug: 315354060
Test: Manual, TaskbarRecentAppsController
Change-Id: Ia593d8896b88822e5cc4df0d8e3221c594e43581