Only update task views when BitmapInfo/text or theme has changed
Currently we update bitmaps of all icons on a recents change, when most
of them may not actually need that. This leads to unnecessary
performance overhead when rebinding tasks. This change minimizes
updates. But we still should update all icons if the shape or monochrome
theme have changed, for instance. Although BitmapInfo may not have
changed in these scenarios, how the BitmapInfo is processed changes.

Flag: com.android.launcher3.enable_recents_in_taskbar
Bug: 315354060
Test: Manual, Perfetto, TaskbarRecentAppsController
Change-Id: Ifd6c6cd147b821b5500e51fbfd73b6c2aef50ec2