Add new directory for compose within taskbar
Copied most compose logic from SysUI in this CL. Everything in this CL should be no-ops until the composeView is created in the follow up. The only files that are NOT the same from SysUI are:
- AmbientCueViewModel.kt
- AmbientCueRepository.kt

In the follow up CLs, the viewModel/repository/interactor are adapted for Launcher. Note: smartspace connection in this CL will be deprecated when ACE integration happens.

Current todos in follow up which should not be Teamfood blocking:
1) frontRunningTask + globallyFocusedTaskId that sysUI implements to keep track of the task that is currently in the foreground. This is used to derive the package name for logging
2) globalFocusedTaskId will have to not be using flows.
3) Migration to use LauncherPrefs instead.

bug:437417091
Test:presubmit
Flag:com.android.systemui.shared.cue_bar_ace_migration
Change-Id: Idc2e0254384c4b0b6745d431a0af9e887639a7e4