Emulate devices through window context instead of test runner.
This approach switches to test parameterization when we want to test
multiple device configurations. It ensures compatibility with tests that
may already need to use a parameterized runner, such as when testing
with flag combiniations.

In many cases, this change reduces the number of devices being emulated
for controller tests where it isn't relevant.

Flag: EXEMPT TEST_ONLY
Bug: 230027385
Test: NexusLauncherRoboTests
Change-Id: Ibb0cdb3de431130ad0705baba38daa5618df60ae