Avoid rebinding Taskbar icons that have not changed.
Anytime we call TaskbarView#updateItems, we rebind every icon whether its backing data changed or not. That means we are regenerating every icon and other parts of BubbleTextView which is inefficient.

This change checks the tag instances of icons to see if they are different. For tag matches, we do not rebind the data. This works for pinned apps because TaskbarModelCallbacks applies any updates to existing pinned apps before committing all items. For recent apps, we use the TaskIconCache to fetch and apply any changes relevant to BubbleTextView (e.g. icon or title).

However, we need a force update mechanism for pinned apps, for when
TaskbarModelCallbacks#bindCompleteModel is invoked. We already do a
similar thing for recents when we need to update them all.

Bug: 315354060
Flag: com.android.launcher3.enable_recents_in_taskbar
Test: manual and TaskbarViewTest
Change-Id: I1105a9ef8e0f845122c66aa18ba83befa9466114