Change logic to add Splitscreen option on taskbar.
Previously, we add this option when we are not in desktop mode - and the
way we check this is via checking how many desktop tasks are currently
open. However in desktop-first devices, it is possible there are no
desktop tasks open yet we still do not want to show splitscreen option.

Instead, changing the logic so that it hides on desktop-first devices,
no matter the desktop tasks visibility.

Bug: 442527551
Test: Manual. Use a desktop-first device and long press on app icon -
no longer see Split screen option.
Flag: EXEMPT bugfix


Change-Id: I580a479ee0a2dc8d0964fde5b34ce3c3c03ceaa7