[Psb2Qsb] Listen to AppStore OnUpdateListener since appstore takes time to update after app is installed.
In case of search engines that doesn't support a widget, we show the searchbox with appInfo icon in it.

When such search engines are selected in setup wizard and app installation takes time, appstore also gets updated later.

OseWidgetView need to listen to AppStore updates, so that the errorView is refreshed when appInfo is ready.

Bug: 458020367
Test: Manual
Flag: com.android.launcher3.enable_qsb_on_hotseat

Change-Id: I9d8f4a1e0bb6f4275ec84b7073bfcd812027fb8e