Breaking circular dependency in DeviceProfile between HotseatProfile and WorkspaceProfile
I think this circular dependency came about by accident because it
wasn't easy to analyse the depencency between the variables but with
careful analysis we can get rid of this circular dependency entirely.
For now I create HotseatProfileInitialValues as an intermidiate step.

That class could have some prunning, not al variables inside of it are
part of the circular dependency but to prevent the change from growing
to large I will prune it in subsequent changes.

Bug: 427811804
Bug: 419264653
Bug: 431261051
Flag: EXEMPT PURE_REFACTOR
Test: DeviceProfileDumpTest
Test: All Image tests
Change-Id: Ifca9efe8b8abca9073e0b551ad089824bdb61fbc