Prevent nav button jump when going home from -1 page
Due to a race condition, if the home button was pressed when on the -1
page, the nav bar would appear to jump when transitioning to home. This was due to conflicting
progress updates between the home animation and the -1 page scrolling. The fix introduces a flag
to temporarily ignore the scroll updates when returning to the home
screen so that the navigation bar doesn't keep translating back and
forth.

Fix: 386114584
Test: In 3 button nav mode, go to -1 page and press the home button. The
nav buttons shouldn't jump up and down when returning to home
Flag: EXEMPT BUGFIX

Change-Id: Ic5786d30491d754e40c397b347dc88c0dbf7255b