Lazily attach base context so that virtual display rule runs first.
We were creating a virtual display for the base context before the rule
even ran, which may lead to it not being released by the rule's
teardown. For instance, if a rule executes before this context and
throws an exemption or has a violated assumption, the virtual display
rule will not run. As a result, the virtual display leaks.

Flag: EXEMPT TEST_ONLY
Fix: 458797022
Test: NexusLauncherTests
Test: NexusLauncherTests:TaskbarWindowSandboxContextTest
Change-Id: I625882bc4877d183e312e8aa48059a667a9992b9