Using internal reference instead of using service connection for assessing TIS
Previously launcher would use service connection to access TaskbarManager (via binding to TIS) and typecast the binder to internal class reference. This was done to simplify the singleton management as android enforces single-instance for services.

But with dagger singleton management is much easier and it avoids extra round-trip to system server. This would also speed up activity creation as the objects are available in the same frame instead of waiting for the service-binding to complete.

Bug: 458439907
Flag: EXEMPT PURE_REFACTOR
Test: Presubmit

Change-Id: I20b6dccc8d0a2b6fb9ea639d8158be2c5e2fdd45