Split on CD: Always show Recents Window on animation start.
When split select animation starts, there is a possibility that
the recents window's visibility is actually false due to hiding
the overview. We should always set the visibility to true via
using #showRecentsWindow (which will also help creating the
window if it does not exist yet).

Bug: 442522952
Test: Manual. Going to Split-select after hiding Overview, and
split select shows now.
Flag: EXEMPT bugfix

Change-Id: Idc6236a6a003e5134321f1ce04b3aaba6a447030