Show recentsView when clicking recents button from splitscreen on CD
This CL fixes the bug when overview is not showing when clicking recents button from splitscreen on CD. The recentsView was hidden by an animation in confirmSplitSelect. This CL makes sure recentsView's CONTENT_ALPHA is set to 1 when entering overview.

Also ensure when RecentsView slide out when exiting to home, its content won't fade out until the last frame.

Bug: 443134857
Test: manual
Flag: EXEMPT BUGFIX
Change-Id: Ia6c16f09b62aa5521e9fae2874e6b6cffcf53d56