Restore bubble bar translation Y on animation interruption
Restore the bubble bar's translation Y if the
`animateBubbleBarForCollapsed()` animation was interrupted, specifically
when an Input Method Editor (IME) is shown.

This restoration is necessary because the bubble bar should not be
stashed on the launcher home or overview screens when the taskbar is in
transient mode. The IME appearing interrupts the animation, leaving the
bubble bar in an incorrect, translated state.

Fixes: 461501572
Fixes: 455422130
Test: BubbleBarViewAnimatorTest
Flag: com.android.wm.shell.enable_bubble_bar
Change-Id: I3ab59268f9373a8b64f03abe279c213ea9d4e861