Update taskbar height after initial bubbles are loaded
When device unfolds, we update taskbar window height. In some cases, the
taskbar window height changes based on bubble bar being visible or not.

At the time of unfold, when we update the taskbar window height, we may
not yet have a signal from shell that there are bubbles. So the taskbar
window does not include extra size for showing bubble bar.

After the shell update does come in, it is an initial state type of
update. And we do not animate this, just show the bubble bar. And this
also does not trigger an update to the taskbar window size. So we do not
know to account for bubble bar.

Add an additional update when processing the initial update. If the
initial update comes in and bubbles are added, we ask the bubble bar
view controller to update taskbar window size to the latest window size.

This ensures that if bubble bar is there, after unfold, the taskbar
window accounts for it and it does not get clipped.

Bug: 458324814
Test: start from folded state with no bubbles, move an app to bubble,
    collapse bubble, unfold device, check that bubble bar shows up
Flag: EXEMPT BUGFIX
Change-Id: Iba6ff514db53a60664f6ce04cf5c964ee838d2b1