taskbar-dnd: Keep the dragged icon being drawn above other siblings
When a taskbar icon is dragged to open the overflow container, the
dragged icon will be drawn below the container. This is mainly
because the ArrowPopup has the elevation set to a positive value,
which is used to draw the shadow below. (Note that other views
like bubbles also have the elevation set). This means that
everything that has lower elevation will be drawn below the
overflow container.

This cl removes the elevation for the overflow container so that
the drawn order is simply defined in the TaskbarDragLayer.

Bug: 447444838
Test: Manual
Flag: com.android.launcher3.enable_taskbar_drag_and_drop
Change-Id: I5d1603af056ef718c3d1a4c97a4d4faceb7e8f7b