Ensure NearestTouchFrame only activates on primary display.
For smaller screens, NearestTouchFrame is used to route touches to
nav buttons to make it easier to tap with limited real estate. Now
that there are secondary displays and simulated secondary displays, the
device size isn't the only factor we should use to see if
NearestTouchFrame should be active. This change adds a check to see if
the current display is the primary display to prevent incorrect touch
routing to the navigation buttons on secondary displays that originate
from a smaller device.

Fix: 447267410
Test: Simulate a secondary display on smaller device like a foldable and connect a mouse. Check
that you are able to launch apps/ your clicks don't get routed to the 3
button nav. Ensure that behavior on the primary display is unaffected.
Flag: EXEMPT BUGFIX

Change-Id: I7349feeb07a31cd52a00799130bb92bf3e5e8f9e