Fix broken launcher state on battery saver mode change
If launcher is destroyed mid-gesture, it will restore its last state. If that state is BACKGROUND_APP, the user will be put in a state where recents view is in full screen and scrollable. Additionally, if launcher is destroyed when a recents animation has already started, TaskAnimationManager won't finish that animation, the user will have to swipe again to finish the animation.

- Updated Launcher state restore logic to either disallow restore absolutely, or with an exemption for theme changes. BACKGROUND_APP state is currently the only state that diables it absolutely.
- Updated TaskAnimationManager to finish started recents animations immediately on launcher destroy
- Removed redundant logs

Flag: EXEMPT BUGFIX
Fixes: 461226985
Test: TaskAnimationManagerTest
Change-Id: I1e17eef55f8eb6cdfc2a2dab0c83325b3036b483