Update TaskbarProfile when recreating taskbar.
This cl fixing issue of folder icon being out of alignment with rest of taskbar icon.

the problem was that for external display we did not recreate device profile to take in any display changes. As solution, we need to recreate device profile for external display always.
problem with primary display was that we were not having the right takbar icon size in Taksbar profile, so even when we update the folder icon size it would be wrong in taksbar eventually.

Test: Manual, Presubmit
Bug: 462078336
Bug: 460526427
Flag: EXEMPT bugfix
Change-Id: I0d9bc07f25445bddaa1e4662436e7973ad0ffc11