Wait for all apps button when showing search EDU
Search EDU tooltip is anchored to all apps button, so it needs all apps
button to be attached to a window in order to be positioned correctly.
This may not be the case when attempting to show EDU tooltip early on,
before taskbar content gets fully initialized.
When attempting to show search EDU, check whether all apps button is
added to the taskbar window - if it's not, add a listener for the button
attached state that attempts showing search EDU when all apps button
gets attached to a window.

Bug: 452149591
Test: Verify that EDU tooltip is correctly positioned when shown just
      after user login on desktop (where taskbar is shown on home).
Flag: EXEMPT BUGFIX
Change-Id: I7ece1fac74cc2f8c2c33bb3da42599d2ab82fbd5