Fix bug for adding item from popup menu in all app
There is a bug when trying to add an item from all apps to a new page
from all apps menu. What happens is that we call removeExtraEmptyScreen
from onDragEnd when we go back to NORMAL state. So there's a race
condition where we remove the new empty screen we added before the item
gets a chance to be added. We also weren't snapping to the page where we
add the item. This fixes the race condition and we go to the page where
the item is to be added.

Fix: 462576490
Flag: EXEMPT BUGFIX
Test: Manual
Change-Id: Icb3d9a0e45a91f425608987d4c4769c73e1b2d0b