Handle widget picker drag and drop after normal state.
When app-specific picker uses this flow, it takes quite a bit of time
for all apps to close and home screen to get to normal state - it
doesn't go into spring loaded state on time. So, we wait for normal
state.

Bug: 436344835
Test: Manual
Flag: com.android.launcher3.enable_app_widget_picker_refactor
Change-Id: I85b09aa8d554facd068cedc72f8d759d8d313a87