AmbientCueRepositoryImpl: Fix leak caused by TaskStackChangeListener
The TaskStackChangeListener was registered as a non-static inner class, causing an implicit strong reference to AmbientCueRepositoryImpl, which in turn strongly referenced the TaskbarActivityContext. Since the listener is registered with the global TaskStackChangeListeners singleton, this prevented the context from being garbage collected, causing a memory leak.

This change refactors the listener into an external class, AmbientCueTaskStackListener, and uses a WeakReference to the AmbientCueRepositoryImpl instance. This should break the strong reference chain, allowing the repository and its associated context to be released when no longer in use.

Looked at Android Studio Profiler to see if there were any leaks and looked at leakCanary logs.

bug: 463269239
bug: 463303093
Test: manually presubmit. LeakCanary
Flag: com.android.systemui.shared.cue_bar_ace_migration
Change-Id: Id32fe190ff25a75ad2c08e461bca2757cab702de