Handle closing ActivityEmbedding during cross-display app launch
When we perform a cross-display app launch for a Task with an
ActivityEmbedding, from a larger to a smaller display, one of the
embedded Activities might not fit on the new display. If that's the case
that Activity should be hidden on the destination display, during the
cross-display app launch animation. This CL hides such Activities.

Flag: com.android.window.flags.enable_cross_displays_app_launch_transition
Bug: 461458558
Test: CrossDisplayMoveTransitionTest
Change-Id: Id1e6b4de7aaec6fb7a944e5b32d4af805b0b5b29